Refactor nftables code to separate test and real impl using an iface
Currently, the nftables code checks whether it’s running in a test or
real environment and changes behavior based on that. This isn’t a clean
design. This PR introduces a proper interface to separate the real and
mock implementations. With this approach, there’s a clear boundary
between test code and production code.

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>